fix windows and macos github action issues
