+++ core/src/Logging.cpp
@@ -30,51 +30,8 @@
 #include <fmt/printf.h>
 #include <fmt/ranges.h>
 
-#include <functional>
-#include <iostream>
-#include <sstream>
-#include <string>
-
 namespace cloudViewer {
 namespace utility {
-
-enum class CV_CORE_LIB_API TextColor {
-    Black = 0,
-    Red = 1,
-    Green = 2,
-    Yellow = 3,
-    Blue = 4,
-    Magenta = 5,
-    Cyan = 6,
-    White = 7
-};
-
-struct CV_CORE_LIB_API Logger::Impl {
-    // The current print function.
-    std::function<void(const std::string &)> print_fcn_;
-
-    // The default print function (that prints to console).
-    static std::function<void(const std::string &)> console_print_fcn_;
-
-    // Verbosity level.
-    VerbosityLevel verbosity_level_;
-
-    // Colorize and reset the color of a string, does not work on Windows,
-    std::string ColorString(const std::string &text,
-                            TextColor text_color,
-                            int highlight_text) const {
-        std::ostringstream msg;
-#ifndef _WIN32
-        msg << fmt::sprintf("%c[%d;%dm", 0x1B, highlight_text,
-                            (int)text_color + 30);
-#endif
-        msg << text;
-#ifndef _WIN32
-        msg << fmt::sprintf("%c[0;m", 0x1B);
-#endif
-        return msg.str();
-    }
-};
 
 std::function<void(const std::string &)> Logger::Impl::console_print_fcn_ =
         [](const std::string &msg) { std::cout << msg << std::endl; };
